Cast label to LongTensor

+++ mutual_info_img_txt/model_utils.py
@@ -139,9 +139,9 @@ class CXRImageDataset(torchvision.datasets.VisionDataset):
         img_id, study_id = self.dataset_metadata.loc[idx, \
             [self.data_key, 'study_id']]
         if self.dataset_metadata.loc[idx,self.disease_label] > 0:
-            labelDisease = 1.0
+            labelDisease = 1
         else:
-            labelDisease=0.0
+            labelDisease=0
 
         if self.cache_images:
             img = self.images[str(idx)]
@@ -154,7 +154,7 @@ class CXRImageDataset(torchvision.datasets.VisionDataset):
 
         img = np.expand_dims(img, axis=0)
 
-        return img, labelDisease
+        return img, labelDisease.type(torch.LongTensor)
 
 # adapted from
 # https://towardsdatascience.com/https-medium-com-chaturangarajapakshe-text-classification-with-transformer-models-d370944b50ca
